Separate auth command classes
