--- Application/DTOs/_Account_/UserDto.cs
@@ -1,10 +1,0 @@
-﻿using Domain.Enums;
-using Mapster;
-
-namespace Application.DTOs._Account_;
-
-public class UserDto 
-{
-    public Guid Id { get; set; }
-    public UserRole Role { get; set; }
-}
