+++ Application/DTOs/_Author_/AuthorDto.cs
@@ -1,4 +1,4 @@
-﻿
+﻿using Application.DTOs._Book_;
 using Domain.Entities;
 using Mapster;
 
@@ -9,7 +9,7 @@ public class AuthorDto : IMapFrom<Author>
     public Guid Id { get; set; }
     public string Name { get; set; } = null!;
     public string Surname { get; set; } = null!;
-    public string Birthdate { get; set; } = null!;
+    public string BirthDate { get; set; } = null!;
     public string Country { get; set; } = null!;
     
 }
