+++ Application/Features/Auth/Commands/RegisterUser/RegisterUserHandler.cs
@@ -1,31 +1,20 @@
-﻿using System.Net.Mail;
-using Application.DTOs._Account_;
+﻿using Application.DTOs._Account_;
 using Application.Interfaces;
 using Application.Interfaces.Repositories;
 using Application.Interfaces.Services;
-using Application.Services;
 using Application.Wrappers;
 using Domain.Entities;
-using Domain.Enums;
-using FluentValidation;
 using Mapster;
 using MediatR;
 
-namespace Application.Features.Auth.Commands;
+namespace Application.Features.Auth.Commands.RegisterUser;
 
-public record RegisterUserCommand : IRequest<Result<TokenPair>>
-{
-    public string Username { get; init; }
-    public string Mail { get; init; }
-    public string Password { get; init; }
-    public UserRole Role { get; init; }
-}
 
 public class RegisterUserHandler(
     IUserRepository repository,
     IJwtService jwtService,
     IUnitOfWork unitOfWork,
-    PasswordService passwordService
+    IPasswordService passwordService
 )
     : IRequestHandler<RegisterUserCommand, Result<TokenPair>>
 {
@@ -43,17 +32,5 @@ public class RegisterUserHandler(
     }
 }
 
-public class RegisterUserValidator : AbstractValidator<RegisterUserCommand>
-{
-    public RegisterUserValidator()
-    {
-        RuleFor(x => x.Role).IsInEnum();
-        RuleFor(x => x.Mail).
-            EmailAddress()
-            .MaximumLength(50)
-            .NotEmpty();
-        RuleFor(x => x.Username).MaximumLength(30).NotEmpty();
-        
-    }
-}
+
